// src/app/models/resena.ts
export class Resena {
    id?: any;
    idpelicula?: number;
    calificacion?: number;
    critica?: string;
    titulo?: string;
    idusuario?: number;
}

export class ResenaFull{
    idresena?: any;
    idpelicula?: number;
    calificacion?: number;
    critica?: string;
    titulo?: string;
    idusuario?: number;
    nombre?: string;
    apellido?:string;
}